Add support for mods with no listeners
In some cases the only reason for a listener is to allow mixin configuration to be registered.  This change allows for a mod creator to specify the mixins configuration filename(s) and Rift will add them to the configuration.  e.g.

{
  "id": "modid",
  "name": "Example mod",
  "authors": [
    "me"
  ],
  "mixins": [
    "mixins.modid.json"
  ]
}

Another benefit for doing this is that some vanilla-only mods can use this without having to mess with their existing build.gradle... Just add a riftmod.json and be done.